backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
beautifulsoup4==4.12.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.0
fastapi==0.104.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.345
langchain-core==0.0.13
langsmith==0.0.92
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.3.7
packaging==23.2
propcache==0.2.1
pyasn1==0.6.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
yarl==1.18.3
